requirements.txt
streamlit
google-generativeai
python-docx
PyPDF2
beautifulsoup4
selenium
requests
langchain
chromedriver-autoinstaller
openai
tiktoken